requirements.txt
streamlit
python-libtorrent
